requirements.txt
# Azure OpenAI API integration
openai>=1.3.0
azure-identity>=1.13.0
azure-keyvault-secrets>=4.7.0

# Web framework for potential API endpoints
fastapi>=0.103.1
uvicorn>=0.23.2

# Development and testing
pytest>=7.4.2
pytest-cov>=4.1.0
black>=23.9.1
flake8>=6.1.0
isort>=5.12.0

# Logging and monitoring
structlog>=23.1.0

# Environment management
python-dotenv>=1.0.0

# Container health checks
requests>=2.31.0

# RCA system dependencies
pydantic>=2.4.2
numpy>=1.24.0
azure-search-documents>=11.4.0
uuid>=1.30 